soc: qcom: smd: Return -EBUSY for a port that exists but is already open
Currently, the smd module returns -ENODEV if an attempt is made
to open a port that is already open. This means that a client
can't tell the difference between the port not existing and the
port being in use by another client.

Fix iterates through the list of currently open ports prior to
opening a new port and checks to see if the port is already open.
If port is already open, fix logs an error message and returns -EBUSY.

Change-Id: I19834ed693f4b58736d73f5acb6fea5fc29beb57
Signed-off-by: Steven Cahail <scahail@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>