soc: qcom: smd: Correct race with reopen ready event
The reopen ready event is sent to clients after they close a channel after
the remote side has opened the channel to indicate when the remote side is
ready for the channel to be reopened.  There exists a race where the remote
side can process the local close state change before we check the remote
state and decide to trigger the reopen ready event.

To address this race, check and cache the remote state before triggering
the local close state change.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>